requirements.txt
apistar==0.5.37
argon2-cffi==18.1.0
blessings==1.7
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
curtsies==0.3.0
greenlet==0.4.13
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.7.0
PyYAML==3.12
requests==2.19.1
six==1.11.0
typing==3.6.4
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.3.1
